+++ server/routes/account.router.js
@@ -391,7 +391,7 @@ router.post('/factoryReset', rejectUnauthenticated, async (req, res) => {
       "reinvest" boolean DEFAULT false,
       "reinvest_ratio" integer DEFAULT 0,
       "max_trade" boolean DEFAULT false,
-      "max_trade_size" integer DEFAULT 0,
+      "max_trade_size" numeric(32,8) DEFAULT 0,
       "profit_reset" timestamp
     );
